Extracts classes in different files
