+++ lib/tictactoe/gui/moves_buffer.rb
@@ -1,0 +1,17 @@
+module Tictactoe
+  module Gui
+    class MovesBuffer
+      def initialize()
+        @moves = []
+      end
+
+      def add(move)
+        @moves.push(move)
+      end
+
+      def get_move!
+        @moves.pop
+      end
+    end
+  end
+end
